+++ src/pretalx/cfp/phrases.py
@@ -5,9 +5,8 @@ from pretalx.common.phrases import Phrases
 
 class CfPPhrases(Phrases, app='cfp'):
 
-    auth_already_requested = _('You already requested a new password within the last 24 hours.')
     auth_password_reset = _(
-        'We will send you an e-mail containing further instructions. If you don\'t '
+        'If we know a user by this email address (who has not requested a password reset in the last 24 hours), we will send you an e-mail containing further instructions. If you don\'t '
         'see the email within the next minutes, check your spam inbox!'
     )
     auth_reset_fail = _(