+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :bug:`-` The password reset form told users if they had already tried to reset their password in the last 24 hours. While this is helpful, it also allows user discovery via password reset. Instead, the user now sees a more generic message.
 - :feature:`-` While filling out the submission form wizard, submitters see a top bar telling them where they are in the submission process.
 - :bug:`-` The event dashboard showed a wrong countdown towards the CfP end while the CfP end was between three and one day away.
 - :feature:`-` The new event dropdown helps you find your event, even if you have access to many events.